Move DatePickers and MultipleSelect
