+++ client/src/components/details/FlightsList.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { fetchFlights } from '../../actions/flightActions';
 import Button from 'material-ui/Button';
-import MultipleSelect from '../MultipleSelect';
-import DatePickers from '../DatePickers';
+import MultipleSelect from './MultipleSelect';
+import DatePickers from './DatePickers';
 import { withStyles } from 'material-ui/styles';
 import PropTypes from 'prop-types';
 import FlightIcon from 'material-ui-icons/Flight';
